requirements.txt
argcomplete==1.9.3
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.8
idna==2.6
mock==2.0.0
python-dateutil==2.6.1
recommonmark==0.4.0
six==1.11.0
smmap2==2.0.3
Sphinx==1.7.4
sphinxcontrib-programoutput==0.11
termcolor==1.1.0
urllib3==1.22
wavefront-api-client==2.3.1
